+++ src/geometry/Circle.js
@@ -5,4 +5,12 @@ export default class Circle {
     this.radius = opts.radius;
     this.diameter = opts.radius * 2;
   }
+
+  area() {
+    return Math.PI * this.radius * this.radius;
+  }
+
+  circumference() {
+    return Math.PI * this.diameter;
+  }
 }