+++ src/geometry/Square.js
@@ -2,6 +2,14 @@ export default class Square {
   constructor(opts) {
     this.x = opts.x;
     this.y = opts.y;
-    this.size = opts.size;
+    this.length = opts.length;
+  }
+
+  area() {
+    return this.length * this.length;
+  }
+
+  perimeter() {
+    return this.length * 4;
   }
 }